Removed ledger parameter from request
The `ledger` parameter is now `ledger_index`, which is optional and defaults to current. This PR removes the `ledger` param from the example requests.

I had opened a PR to simply change `ledger` to `ledger_index`, however the conventions of the API docs seem to only use required parameters, so I closed that PR and created this one instead.