Removed ledger param from account_objects
Similar to PR #409 except the deprecated `ledger` param was used in the websocket request example, while `ledger_index` was used in the JSON example.

This PR removes those parameters on both examples. I also removed the word `Optional` from the `Type` column since the convention seems to be to indicate that in the description field.

Note: The Commandline request example has `[ledger_index]`, however the syntax suggests it's an optional argument and not an flagged option (i.e. `--ledger_index=current`), so it seems proper to leave that as is.