Remove forum links
The old Ripple forum is down indefinitely for security reasons.
Until the contents can be properly archived in a secure (and read-only)
way, forum links won't work.

There was a Wayback Machine archive for the ledgers thing, so I changed
it to use that. I could not find an equivalent for the drop naming thread,
so I removed that hyperlink.