`token_sequence` must be incremented in backups
Docs indicate that validators should backup `validator-keys.json`, and not edit the backup. This causes problems when validator operators generate new validation tokens using backups of the `validator-keys.json` file. Specifically:
1. Validator operator copies backup of `validator-keys.json` to from backup location to validator
2. Operator generates a new token on validator
3. Operator deletes `validator-keys.json` from the validator

If the validator operator does not then increment the "token_sequence" in the backed up `validator-keys.json`, the next time the operator repeats the above three steps, their validations will be ignored by the network, as the resultant token will have a sequence number that is <= the last observed number.