Disable language selector
The Japanese translated version of xrpl.org has been falling behind.
Since it's a worse experience to see a page that's mostly untranslated,
which is the case for many of the new/updated page, and traffic to the
Japanese version of the site is <1% of total traffic, we are temporarily
disabling the language selector so that we don't direct people to the
largely-untranslated pages. The existing URLs, including pages that have
been fully translated, continue to work.