Add DestinationNode to the code sample
DestinationNode is listed in the parameter descriptions, but missing from sample code. I've added the entry, copying the OwnerNode value for now. I'm assuming that the value is a placeholder with the correct number of digits. If we can put in a more realistic value, that would probably be better.