Reword content articles to use "tokens"
This replaces the old wording of "issued currencies" and introduces
broader usage of the term "stablecoins" to more closely match the
terminology in use by the wider industry.

I've also added a draft "Common Misunderstandings about Freezes" page so
that the Freeze page doesn't have to protest quite so much, and written
a very brief word on tokens' use for community credit.

This commit only covers the Concepts section, in English, and likely
leaves some links to the old URLs for the renamed pages.