Update federated-sidechains.md
A typo, it's supposed to be `Federated Sidechain` as it allows you to transfer assets in-between 2 networks.

The 1st one says it's a 'blockchain based on the XRPL technology' meaning it's a separate network that isn't connected to the mainnet. Whereas the 2nd one includes proxy assets and door accounts which are both included in *federated* sidechains.

Since there are 2 terms of the word `sidechains`, it's best to not confuse the reader by including the word `federated` on sidechains that are able to transfer assets in-between each other.