Fixed Public Key Generation Description
Hi there,
To use the PaymentChannelCreate method, you need to create a public key. However, I believe the description is now incorrect/not clear enough, so I changed it.

I used the PaymentChannelCreate method in the Testnet, with a Public key that was 66 Characters long, derived from the sources' private key (secret), and hashed it with the SHA-512 algorithm. Then I took the first 66 Characters and used them as the public key, and the PaymentChannel was created. 

Why 66 Characters?
In the Documentation for PaymentChannelCreate (https://xrpl.org/paymentchannelcreate.html) you are using this Public Key: "32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A", which is 66 characters long, and from what I believe is derived from your secret key.

The documentation may be improved with my description to clarify how to generate the PublicKey/how it is generated.