Fix reports of broken links, disable funding page
Images whose content is defined in CSS have "data:," as their src
attribute because it's technically required and some tools may complain
about it missing, but the link checker complains about the data URI, so
this change makes the link checker skip these.

Also, the funding page is intended to go along with a Bounties program
that has not yet launched, so this disables it for now.