Improve install-rippled-on-ubuntu.md
`mv` will rename the file (instead of moving it into a destination directory) if the destination directory does not exist. By providing the `/`, we are being explicit that we want the file to be moved.

If the earlier `mkdir` command was skipped or failed, we want `mv`  to report an error, instead of silently renaming the key file to `keyrings`, which would not make sense.

Supplying a command to `gpg` prevents the warning which was previously shown in the gpg output.