Move/rename indexes and some other files
This makes no difference to how the site is built under Dactyl (other
than where the "Edit Page" links go) but changes the URLs that will
be used for the pages after the migration to Redocly.

In addition to renaming index pages to index.md, I updated the style
guide to reflect the updated conventions, and moved a couple files that
were not in the correct folders for their place in the nav hierarchy.