Remove `currentpage.md` variables
As part of the Redocly migration, I'm removing syntax that isn't
compatible with the new portal. In many of these cases, the existing
if/else statements weren't working as intended anyway because they
hadn't been updated after the files had been renamed.

In the case of the Client Libraries page, I've removed mention of
editing the page to contribute client libraries for now; when someone
adds guidelines (issue #2266) it would make sense to add a link to those
guidelines in that place instead.