Scripting for all-in-one link replacement
- Add tool/migrate_links.sh as a one-stop conversion script for links.
- Enable the update_links filter in dactyl config but make it inactive
  unless you pass the appropriate vars
- Hack include_svg script to assume content/img instead of img

The script is meant to be run from the repo top. It replaces links
in-place. Unless this is the final migration phase, the results of
running the migration script should be committed in a separate commit
with "[DROP]" in the commit message so it can be re-run after rebasing
on the latest master branch.