Accept date in badges. Fix errors in build.
Our guidelines suggest adding a date when using the badge markdoc
component, but until now the component's schema didn't have a defintion
for the date parameter, so it was reporting errors.

The date doesn't affect how the badge is displayed on the site, but
we can use it to see when a 'New in' badge is old enough to no longer be
relevant (typically >2yrs) so we can remove it.